将logo放在SPIFFS中
