+++ GRBL/src/Config.h
@@ -99,7 +99,7 @@ const int MAX_N_AXIS = 6;
 /* CONFIGURE_EYECATCH_BEGIN (DO NOT MODIFY THIS LINE) */
 //#define ENABLE_BLUETOOTH                  /* enable bluetooth  */ 
 #define ENABLE_SD_CARD                      /* enable use of SD Card to run jobs */
-#define ENABLE_WIFI                         /* enable wifi */
+// #define ENABLE_WIFI                         /* enable wifi */
 #define ENABLE_TFT                          /* enable tft lcd for lvgl ui @ V8.3 */
 
 #if defined(ENABLE_WIFI) || defined(ENABLE_BLUETOOTH)
