+++ GRBL/src/SDCard.cpp
@@ -24,13 +24,11 @@
 
 SDCard mysdcard;
 
-// File                       myFile;
-// bool                       SD_ready_next = false;  // Grbl has processed a line and is waiting for another
 uint8_t                    SD_client     = CLIENT_SERIAL;
 WebUI::AuthenticationLevel SD_auth_level = WebUI::AuthenticationLevel::LEVEL_GUEST;
 uint32_t                   sd_current_line_number;     // stores the most recent line number read from the SD
 static char                comment[LINE_BUFFER_SIZE];  // Line to be executed. Zero-terminated.
-SDState sd_state =          SDState::Idle;
+SDState sd_state =         SDState::Idle;
 
 #define USE_HSPI_FOR_SD 1
 #ifdef USE_HSPI_FOR_SD
